renamed main. hanabi -> hanabi, game
# Conflicts:
#	hanabi.py
#	players/oracle.py
